VIRUSBreakend: using default GRIDSS jvm heap size since Kraken2 needs more memory anyway

+++ scripts/virusbreakend_manuscript/scripts/cleanup_gce.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 while true ; do
 	echo "$(date) Awake"
-	gcloud compute instances delete -q $( gcloud compute instances list --filter="name~virusbreakend" | grep TERMINATED | cut -f 1 -d ' ' )
+	tokill=$( gcloud compute instances list --filter="name~virusbreakend" | grep TERMINATED | cut -f 1 -d ' ' )
+	if [[ "$tokill" != "" ]] ; then
+		gcloud compute instances delete -q $( gcloud compute instances list --filter="name~virusbreakend" | grep TERMINATED | cut -f 1 -d ' ' )
+	else
+		echo "Nothing to kill"
+	fi
 	sleep 180
 done
